refactor : 패키지 명 수정, 패키지 위치 수정, swaager API 설명 추가
- 패키지명 소문자로 수정
- 패키지 위치 변경
- controller,dto swagger 설명 추가
