+++ src/main/java/com/ureca/gate/member/controller/MemberRedirectController.java
@@ -1,6 +1,8 @@
 package com.ureca.gate.member.controller;
 
 import com.ureca.gate.global.config.security.OauthConfig;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,6 +13,7 @@ import org.springframework.web.bind.annotation.RestController;
 import java.io.IOException;
 
 
+@Tag(name = "Member API", description = "회원 API")
 @RestController
 @RequiredArgsConstructor
 @Slf4j
@@ -18,6 +21,7 @@ import java.io.IOException;
 public class MemberRedirectController {
     private final OauthConfig oauthConfig;
 
+    @Operation(summary = "소셜 로그인 API (카카오)", description = "소셜 회원/로그인 ")
     @GetMapping("/social-login/kakao")
     public void initiateKakaoLogin(HttpServletResponse response) throws IOException {
         String authUrl = "https://kauth.kakao.com/oauth/authorize?" +
