+++ src/main/java/com/ureca/gate/member/controller/request/NicknameCheckRequest.java
@@ -1,8 +1,10 @@
 package com.ureca.gate.member.controller.request;
 
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 
 @Getter
 public class NicknameCheckRequest {
+    @Schema(description = "닉네임", example = "쥬쥬")
     private String nickName;
 }
