--- src/main/java/com/ureca/gate/member/domain/Enum/Gender.java
@@ -1,5 +1,0 @@
-package com.ureca.gate.member.domain.Enum;
-
-public enum Gender {
-    MALE, FEMALE
-}
