+++ src/main/java/com/ureca/gate/member/domain/enumeration/Gender.java
@@ -1,0 +1,6 @@
+package com.ureca.gate.member.domain.enumeration;
+
+public enum Gender {
+    MALE, FEMALE
+
+}
