+++ src/main/java/com/ureca/gate/member/domain/enumeration/Status.java
@@ -1,0 +1,5 @@
+package com.ureca.gate.member.domain.enumeration;
+
+public enum Status {
+    ACTIVE,NONACTIVE
+}
