+++ src/main/java/com/ureca/gate/member/infrastructure/jpaadapter/MemberJpaRepository.java
@@ -1,7 +1,7 @@
-package com.ureca.gate.member.infrastructure.jpaAdapter;
+package com.ureca.gate.member.infrastructure.jpaadapter;
 
 
-import com.ureca.gate.member.infrastructure.jpaAdapter.Entitiy.MemberEntity;
+import com.ureca.gate.member.infrastructure.jpaadapter.entity.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
